# packages/contract/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "REST API layanan inspeksi & sampling lapangan — Go + Gin dengan struktur modular ala NestJS.\nAutentikasi di luar cakupan PoC: seluruh endpoint terbuka dan peran dipilih di frontend.",
        "title": "Verifield API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/orders": {
            "get": {
                "description": "Klien otomatis hanya melihat order milik perusahaannya sendiri, apa pun isi company_id.",
                "tags": [
                    "job-orders"
                ],
                "summary": "Daftar job order",
                "parameters": [
                    {
                        "description": "Id pengguna yang bertindak",
                        "name": "X-Actor-Id",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Halaman",
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "description": "Jumlah data per halaman",
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "description": "Cari ref/objek/lokasi/kota",
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Saring status",
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "requested",
                                "assigned",
                                "on_the_way",
                                "on_site",
                                "in_progress",
                                "completed",
                                "failed",
                                "cancelled"
                            ]
                        }
                    },
                    {
                        "description": "Saring perusahaan",
                        "name": "company_id",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Saring inspektor",
                        "name": "inspector_id",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Saringan layar koordinator",
                        "name": "attention",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "penugasan",
                                "pembatalan",
                                "basi",
                                "terlambat"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/verifield-be_internal_common_response.Envelope"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/verifield-be_internal_modules_joborder_dto.JobOrderResponse"
                                                    }
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "job-orders"
                ],
                "summary": "Buat permintaan inspeksi baru",
                "parameters": [
                    {
                        "description": "Id pengguna yang bertindak",
                        "name": "X-Actor-Id",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/verifield-be_internal_modules_joborder_dto.CreateJobOrderDTO"
                            }
                        }
                    },
                    "description": "Data permintaan",
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/verifield-be_internal_common_response.Envelope"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/verifield-be_internal_modules_joborder_dto.JobOrderResponse"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/verifield-be_internal_common_response.Envelope"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/orders/{id}": {
            "get": {
                "tags": [
                    "job-orders"
                ],
                "summary": "Detail job order beserta riwayat status",
                "parameters": [
                    {
                        "description": "Id pengguna yang bertindak",
                        "name": "X-Actor-Id",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Job order ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/verifield-be_internal_common_response.Envelope"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/verifield-be_internal_modules_joborder_dto.JobOrderResponse"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/verifield-be_internal_common_response.Envelope"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/orders/{id}/assign": {
            "post": {
                "description": "expected_version wajib; bila order sudah berubah, permintaan ditolak dengan 409 (keputusan B-09).",
                "tags": [
                    "job-orders"
                ],
                "summary": "Tugaskan inspektor",
                "parameters": [
                    {
                        "description": "Id pengguna yang bertindak",
                        "name": "X-Actor-Id",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Job order ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/verifield-be_internal_modules_joborder_dto.AssignInspectorDTO"
                            }
                        }
                    },
                    "description": "Inspektor dan versi yang diharapkan",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/verifield-be_internal_common_response.Envelope"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/verifield-be_internal_modules_joborder_dto.JobOrderResponse"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/verifield-be_internal_common_response.Envelope"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/orders/{id}/cancel": {
            "post": {
                "description": "Hasilnya bergantung pada tahap dan peran. Setelah pekerjaan dimulai, pembatalan oleh klien menjadi permintaan yang menunggu koordinator (keputusan B-05).",
                "tags": [
                    "job-orders"
                ],
                "summary": "Batalkan atau ajukan pembatalan",
                "parameters": [
                    {
                        "description": "Id pengguna yang bertindak",
                        "name": "X-Actor-Id",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Job order ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/verifield-be_internal_modules_joborder_dto.CancelJobOrderDTO"
                            }
                        }
                    },
                    "description": "Alasan pembatalan",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/verifield-be_internal_common_response.Envelope"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/verifield-be_internal_modules_joborder_dto.CancelResult"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/verifield-be_internal_common_response.Envelope"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/orders/{id}/cancellations/{requestId}/decide": {
            "post": {
                "tags": [
                    "job-orders"
                ],
                "summary": "Putuskan permintaan pembatalan",
                "parameters": [
                    {
                        "description": "Id pengguna yang bertindak",
                        "name": "X-Actor-Id",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Job order ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Cancellation request ID (UUID)",
                        "name": "requestId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/verifield-be_internal_modules_joborder_dto.DecideCancellationDTO"
                            }
                        }
                    },
                    "description": "Keputusan koordinator",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/verifield-be_internal_common_response.Envelope"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/verifield-be_internal_modules_joborder_dto.JobOrderResponse"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/orders/{id}/corrections": {
            "post": {
                "description": "Boleh mundur ke tahap sebelumnya, wajib beralasan, dan tercatat sebagai entri baru — riwayat tidak pernah ditimpa (keputusan B-06).",
                "tags": [
                    "job-orders"
                ],
                "summary": "Koreksi status oleh koordinator",
                "parameters": [
                    {
                        "description": "Id pengguna yang bertindak",
                        "name": "X-Actor-Id",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Job order ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/verifield-be_internal_modules_joborder_dto.CorrectStatusDTO"
                            }
                        }
                    },
                    "description": "Status hasil koreksi beserta alasannya",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/verifield-be_internal_common_response.Envelope"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/verifield-be_internal_modules_joborder_dto.JobOrderResponse"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "components": {
        "schemas": {
...
            "verifield-be_internal_common_response.Envelope": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string"
                    },
                    "data": {},
                    "errors": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/verifield-be_internal_common_apperror.FieldError"
                        }
                    },
                    "message": {
                        "type": "string"
                    },
                    "meta": {
                        "$ref": "#/components/schemas/verifield-be_internal_common_response.Meta"
                    },
                    "success": {
                        "type": "boolean"
                    }
                }
            },
...
            "verifield-be_internal_modules_joborder_dto.AssignInspectorDTO": {
                "type": "object",
                "required": [
                    "expected_version",
                    "inspector_id"
                ],
                "properties": {
                    "expected_version": {
                        "type": "integer",
                        "minimum": 1,
                        "example": 1
                    },
                    "inspector_id": {
                        "type": "string"
                    }
                }
            },
            "verifield-be_internal_modules_joborder_dto.CancelJobOrderDTO": {
                "type": "object",
                "required": [
                    "reason"
                ],
                "properties": {
                    "expected_version": {
                        "description": "Kosong berarti pemanggil tidak peduli pada versi. Diisi oleh koordinator\nyang bertindak dari layar yang mungkin sudah basi.",
                        "type": "integer",
                        "minimum": 1
                    },
                    "reason": {
                        "type": "string",
                        "maxLength": 500,
                        "minLength": 3,
                        "example": "Jadwal pengapalan dimajukan"
                    }
                }
            },
...
        }
    }
}